# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  template_internal_http_datatransfers_requests.AddEIDSignerRequest:
    properties:
      role:
        maxLength: 100
        type: string
      signer_reg_no:
        maxLength: 20
        minLength: 8
        type: string
    required:
    - signer_reg_no
    type: object
...
  template_internal_http_datatransfers_requests.EIDOrgRegisterRequest:
    properties:
      reg_no:
        maxLength: 16
        minLength: 4
        type: string
    required:
    - reg_no
    type: object
...
  template_internal_http_datatransfers_responses.EIDActivityCounts:
    properties:
      authentication:
        type: integer
      signature:
        type: integer
    type: object
  template_internal_http_datatransfers_responses.EIDActivityItem:
    properties:
      doc_text:
        type: string
      extra:
        additionalProperties: {}
        description: activity service-ийн нэмэлт талбарууд
        type: object
      flow:
        type: string
      outcome:
        type: string
      session_id:
        type: string
      timestamp:
        type: string
    type: object
  template_internal_http_datatransfers_responses.EIDActivityResponse:
    properties:
      counts:
        $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDActivityCounts'
      sessions:
        items:
          $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDActivityItem'
        type: array
      total:
        type: integer
    type: object
  template_internal_http_datatransfers_responses.EIDCertCounts:
    properties:
      expired:
        type: integer
      revoked:
        type: integer
      suspended:
        type: integer
      total:
        type: integer
      valid:
        type: integer
    type: object
...
  template_internal_http_datatransfers_responses.EIDCertificatesResponse:
    properties:
      certificates:
        items:
          $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDPersonCert'
        type: array
      counts:
        $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDCertCounts'
    type: object
  template_internal_http_datatransfers_responses.EIDDevicesResponse:
    properties:
      active_count:
        type: integer
      devices:
        items:
          $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDPersonDevice'
        type: array
      total:
        type: integer
    type: object
...
  template_internal_http_datatransfers_responses.EIDPersonCert:
    properties:
      certificate_level:
        type: string
      document_number:
        type: string
      issuer_dn:
        type: string
      not_after:
        type: string
      not_before:
        type: string
      serial_number:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  template_internal_http_datatransfers_responses.EIDPersonDevice:
    properties:
      active:
        type: boolean
      deactivated_at:
        type: string
      document_number:
        type: string
      enrolled_at:
        type: string
      extra:
        additionalProperties: {}
        description: upstream-ийн нэмэлт талбарууд
        type: object
      platform:
        type: string
    type: object
...
  template_internal_http_datatransfers_responses.EIDSummaryResponse:
    properties:
      activity:
        $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDActivityCounts'
      certificates:
        $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDCertCounts'
      devices_active:
        type: integer
      devices_total:
        type: integer
      given_name:
        type: string
      representation_count:
        type: integer
      surname:
        type: string
    type: object
...
  template_internal_http_datatransfers_responses.OrgPendingConfirmationResponse:
    properties:
      session_id:
        type: string
      signer_etsi:
        type: string
      signer_reg_no:
        type: string
    type: object
  template_internal_http_datatransfers_responses.OrgRepresentationResponse:
    properties:
      org_etsi:
        type: string
      org_name:
        type: string
      org_name_en:
        type: string
      org_register:
        type: string
      right_type:
        type: string
      role:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
...
  template_internal_http_datatransfers_responses.OrgSignerResponse:
    properties:
      name:
        type: string
      name_en:
        type: string
      person_etsi:
        type: string
      reg_no:
        type: string
      right_type:
        type: string
      role:
        type: string
      self:
        type: boolean
      source:
        type: string
      status:
        description: ACTIVE | PENDING (sign-push баталгаажуулалт)
        type: string
    type: object
  template_internal_http_datatransfers_responses.OrgSignersResultResponse:
    properties:
      pending_confirmation:
        $ref: '#/definitions/template_internal_http_datatransfers_responses.OrgPendingConfirmationResponse'
      signers:
        items:
          $ref: '#/definitions/template_internal_http_datatransfers_responses.OrgSignerResponse'
        type: array
    type: object
...
  template_internal_http_handlers_v1.BaseResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Government Developer Portal V3.0
    url: https://github.com/snykk/go-rest-boilerplate
  description: chi (net/http) + pgx (PostgreSQL) + Redis дээр суурилсан Clean Architecture
    бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib
    Fikri)-ээс үүсэлтэй; HTTP давхаргыг chi, өгөгдлийн давхаргыг pgx руу хөрвүүлсэн.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/snykk/go-rest-boilerplate
  title: Government Developer Portal V3.0 API
  version: "1.0"
paths:
...
  /users/me/eid/activity:
    get:
      parameters:
      - description: хуудасны хэмжээ (default 20)
        in: query
        name: limit
        type: integer
      - description: эхлэх байрлал
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDActivityResponse'
              type: object
      security:
      - BearerAuth: []
      summary: eID auth/sign түүх + тоо (RP-scoped)
      tags:
      - users
  /users/me/eid/certificates:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDCertificatesResponse'
              type: object
      security:
      - BearerAuth: []
      summary: eID гэрчилгээний жагсаалт + тоо
      tags:
      - users
  /users/me/eid/devices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDDevicesResponse'
              type: object
      security:
      - BearerAuth: []
      summary: eID холбоотой төхөөрөмжүүд
      tags:
      - users
  /users/me/eid/organizations:
    get:
      description: Нэвтэрсэн eID хэрэглэгчийн eidmongolia.mn-д бүртгэлтэй, төлөөлж
        чадах байгууллагуудыг буцаана. eID-ээр нэвтрээгүй хэрэглэгчид хоосон жагсаалт.
      produces:
      - application/json
      responses:
        "200":
          description: Representations
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/template_internal_http_datatransfers_responses.OrgRepresentationResponse'
                  type: array
              type: object
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "500":
          description: eID provider error
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Төлөөлдөг байгууллагууд (eID)
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Улсын бүртгэлээс (XYP) байгууллагыг регистрийн дугаараар хайж,
        нэвтэрсэн иргэнийг (eID РД нь тухайн байгууллагын захирал/үүсгэн байгуулагч/хувь
        эзэмшигч бол) eidmongolia.mn-д төлөөлөл болгон холбоно. Иргэний бүх төлөөлдөг
        байгууллагыг буцаана.
      parameters:
      - description: Байгууллагын регистрийн дугаар
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/template_internal_http_datatransfers_requests.EIDOrgRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Representations
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/template_internal_http_datatransfers_responses.OrgRepresentationResponse'
                  type: array
              type: object
        "400":
          description: Invalid body
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "403":
          description: Not authorized to represent this organization
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Байгууллага холбох (eID)
      tags:
      - users
  /users/me/eid/organizations/{regNo}:
    delete:
      description: Нэвтэрсэн иргэн өөрийн төлөөлдөг байгууллагын (регистрийн дугаараар)
        холбоосыг цуцлана. Иргэний үлдсэн төлөөлдөг байгууллагыг буцаана.
      parameters:
      - description: Байгууллагын регистрийн дугаар
        in: path
        name: regNo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/template_internal_http_datatransfers_responses.OrgRepresentationResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Байгууллага салгах (eID)
      tags:
      - users
  /users/me/eid/organizations/{regNo}/signers:
    delete:
      description: Нэвтэрсэн иргэний төлөөлдөг байгууллагаас гарын үсэг зурагчийг
        (регистрийн дугаараар) хасна.
      parameters:
      - description: Байгууллагын регистрийн дугаар
        in: path
        name: regNo
        required: true
        type: string
      - description: Хасах гарын үсэг зурагчийн РД
        in: query
        name: signer
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/template_internal_http_datatransfers_responses.OrgSignerResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Гарын үсэг зурагч хасах (eID)
      tags:
      - users
    get:
      description: Нэвтэрсэн иргэний төлөөлдөг байгууллагын гарын үсэг зурах эрхтэй
        хүмүүс. Иргэн тухайн байгууллагын төлөөлөгч байх ёстой.
      parameters:
      - description: Байгууллагын регистрийн дугаар
        in: path
        name: regNo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/template_internal_http_datatransfers_responses.OrgSignerResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Байгууллагын гарын үсэг зурагчид (eID)
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Нэвтэрсэн иргэний төлөөлдөг байгууллагад өөр eID иргэнийг (регистрийн
        дугаараар) гарын үсэг зурах эрхтэй болгож нэмнэ.
      parameters:
      - description: Байгууллагын регистрийн дугаар
        in: path
        name: regNo
        required: true
        type: string
      - description: Гарын үсэг зурагчийн РД + үүрэг
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/template_internal_http_datatransfers_requests.AddEIDSignerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Signers + pending confirmation
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/template_internal_http_datatransfers_responses.OrgSignersResultResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Гарын үсэг зурагч нэмэх (eID)
      tags:
      - users
  /users/me/eid/organizations/{regNo}/signers/resend:
    post:
      description: Баталгаажаагүй (PENDING) гарын үсэг зурагч руу eID sign-push баталгаажуулах
        хүсэлтийг дахин илгээнэ.
      parameters:
      - description: Байгууллагын регистрийн дугаар
        in: path
        name: regNo
        required: true
        type: string
      - description: Гарын үсэг зурагчийн РД
        in: query
        name: signer
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/template_internal_http_datatransfers_responses.OrgSignersResultResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Баталгаажуулах хүсэлт дахин илгээх (eID)
      tags:
      - users
  /users/me/eid/summary:
    get:
      description: Нэвтэрсэн иргэний гэрчилгээ/auth-sign/төхөөрөмж/байгууллагын нэгдсэн
        тоолол (PKI_READ эрхтэй RP). Эрхгүй бол 403.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDSummaryResponse'
              type: object
      security:
      - BearerAuth: []
      summary: eID PKI самбарын нэгдсэн тоо
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: /auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен
      (access token).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"